release: v0.2.0 — 清除无用代码、优化性能、修复 clippy 警告
- 删除字体切换功能及 font_list.rs（注册表字体枚举/查找全部移除）
- setup_fonts 简化为一次性初始化，移除 hash 变更检测开销
- 移除 translate.rs 中全局圆角清零（ctx.style_mut_of）导致的按钮/输入框圆角丢失
- 移除历史窗口记录数描述行
- 修复全部 14 个 clippy 警告：derivable_impls、redundant_closure、unnecessary_cast、manual_find、chunks_exact_to_as_chunks、upper_case_acronyms
- eframe features 精简为 default_fonts + glow
- 版本号升级至 0.2.0

+++ src/ui/components.rs
@@ -62,7 +62,7 @@ pub fn hover_anim_alpha(ui: &mut egui::Ui, resp: &egui::Response) -> f32 {
 
 /// 持久化动画值，每帧向 target 逼近（帧率无关）
 pub fn anim_towards(ui: &mut egui::Ui, id: egui::Id, target: f32, speed: f32) -> f32 {
-    let dt = ui.input(|i| i.unstable_dt).min(0.1) as f32;
+    let dt = ui.input(|i| i.unstable_dt).min(0.1);
     let current = ui.memory(|m| m.data.get_temp::<f32>(id).unwrap_or(0.0));
     let new = current + (target - current) * (1.0 - (1.0 - speed).powf(dt * 60.0));
     ui.memory_mut(|m| m.data.insert_temp(id, new));
